On the ev2-export branch: Bring up-to-date with trunk.
git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/ev2-export@1333596 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/libsvn_ra_serf/util.c
@@ -1326,7 +1326,7 @@ inject_to_parser(svn_ra_serf__xml_parser_t *ctx,
 {
   int xml_status;
 
-  xml_status = XML_Parse(ctx->xmlp, data, len, 0);
+  xml_status = XML_Parse(ctx->xmlp, data, (int) len, 0);
   if (xml_status == XML_STATUS_ERROR && !ctx->ignore_errors)
     {
       if (sl == NULL)
